requirements.txt
pytorch
torchaudio
speechbrain
librosa
tensorboard
tgt
ctc-segmentation
kaldiio
